pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fino"
version = "0.1.0"
description = "FiNo — A CLI tool for secure, decentralized file sharing using IPFS + Nostr"
authors = [{ name = "arnispen" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"

dependencies = [
  "cryptography",
  "requests",
  "websocket-client",
  "pynostr",
  "python-dotenv",
  "ecdsa",
  "bech32"
]

[project.scripts]
fino = "fino.fino:main"
